+++ lib/methods/base.py
@@ -69,6 +69,30 @@ class BaseMethod(object):
     pass
 
 
+  def expandVariablesImpl(self, prefix, variables, result):
+    for key in variables:
+      if isinstance(variables[key], dict):
+        self.expandVariablesImpl(prefix + "." + key, variables[key], result)
+      elif isinstance(variables[key], list):
+        pass # lists are not supported.
+      else:
+        result["%" + prefix + "." + key + "%"] = str(variables[key])
+
+  def expandVariables(self, variables):
+    results = {}
+    for key in variables:
+      self.expandVariablesImpl(key, variables[key], results)
+
+    return results
+
+  def expandCommands(self, commands, replacements):
+    parsed_commands = []
+    pattern = re.compile('|'.join(re.escape(key) for key in replacements.keys()))
+    for line in commands:
+      result = pattern.sub(lambda x: replacements[x.group()], line)
+      parsed_commands.append(result)
+
+    return parsed_commands
   def addPasswordToFabricCache(self, user, host, port, password, **kwargs):
     host_string = join_host_strings(user, host, port)
     env.passwords[host_string] = password
@@ -116,9 +140,6 @@ class BaseMethod(object):
       return False
     return file[0]
 
-
-
-
   def runLocally(self, config):
     return LocallyContext(self, config)
 
@@ -127,16 +148,31 @@ class BaseMethod(object):
     self.setExecutables(config)
 
   def setExecutables(self, config):
-    self.executables = config['executables']
+
+    self.executables = {}
+    replacements = self.expandVariables({ "host": config })
+    for key, command in config['executables'].iteritems():
+      cmds = [ command ];
+      cmds = self.expandCommands(cmds, replacements)
+      self.executables[key] = cmds[0]
+
+
 
   def cd(self, path):
     # print red('cd: %d %s'% (self.run_locally,  path))
     return lcd(path) if self.run_locally else cd(path)
 
-  def expandCommand(self, cmd):
+  def expandCommand(self, in_cmd):
+    cmd = in_cmd
     if len(self.executables) > 0:
-      pattern = re.compile('|'.join(re.escape("#!" + key) for key in self.executables.keys()))
-      cmd = pattern.sub(lambda x: self.executables[x.group()[2:]], cmd)
+      pattern = re.compile('|'.join(re.escape("#!" + key)+"\s" for key in self.executables.keys()))
+      cmd = pattern.sub(lambda x: self.executables[x.group()[2:-1]] + ' ', cmd)
+      if cmd.find('%arguments%') >= 0:
+        arr = in_cmd.split(' ')
+        command = arr.pop(0)
+        arguments = ' '.join(arr)
+        command = command.replace('#!', '');
+        cmd = self.executables[command].replace('%arguments%', arguments)
 
     return cmd